chore: replace legacy web service with Insights client instance
