+++ Tests/InstantSearchInsightsTests/Helpers/XCTest+Codable.swift
@@ -17,7 +17,6 @@ func AssertEncodeDecode<T: Codable>(_ value: T, _ rawValue: JSON, file: StaticSt
 func AssertDecode<T: Codable & Equatable>(_ input: JSON, expected: T, file: StaticString = #file, line: UInt = #line) throws {
 
   let encoder = JSONEncoder()
-//  encoder.dateEncodingStrategy = .swiftAPIClient
   let data = try encoder.encode(input)
 
   let jsonDecoder = JSONDecoder()
@@ -30,11 +29,9 @@ func AssertDecode<T: Codable & Equatable>(_ input: JSON, expected: T, file: Stat
 func AssertDecode<T: Codable>(_ input: JSON, expected: T, file: StaticString = #file, line: UInt = #line) throws {
 
   let encoder = JSONEncoder()
-//  encoder.dateEncodingStrategy = .swiftAPIClient
   let data = try encoder.encode(input)
 
   let jsonDecoder = JSONDecoder()
-//  jsonDecoder.dateDecodingStrategy = .swiftAPIClient
   let decoded = try jsonDecoder.decode(T.self, from: data)
   
   let decodedJSON = try JSON(decoded)
@@ -43,19 +40,12 @@ func AssertDecode<T: Codable>(_ input: JSON, expected: T, file: StaticString = #
   XCTAssertEqual(expectedJSON, decodedJSON, file: (file), line: line)
 }
 
-//@discardableResult func AssertDecode<T: Decodable>(jsonFilename filename: String, expected: T.Type, file: StaticString = #file, line: UInt = #line) throws -> T {
-//  let data = try Data(filename: filename)
-//  return try JSONDecoder().decode(T.self, from: data)
-//}
-
 func AssertEncode<T: Encodable>(_ value: T, expected: JSON, file: StaticString = #file, line: UInt = #line) throws {
 
   let encoder = JSONEncoder()
-//  encoder.dateEncodingStrategy = .swiftAPIClient
   let valueData = try encoder.encode(value)
 
   let jsonDecoder = JSONDecoder()
-//  jsonDecoder.dateDecodingStrategy = .swiftAPIClient
   let jsonFromValue = try jsonDecoder.decode(JSON.self, from: valueData)
 
   XCTAssertEqual(jsonFromValue, expected, file: (file), line: line)
@@ -63,7 +53,6 @@ func AssertEncode<T: Encodable>(_ value: T, expected: JSON, file: StaticString =
 
 func AssertEquallyEncoded<A: Encodable, B: Encodable>(_ l: A, _ r: B, file: StaticString = #file, line: UInt = #line) throws {
   let encoder = JSONEncoder()
-//  encoder.dateEncodingStrategy = .swiftAPIClient
   let lData = try encoder.encode(l)
   let rData = try encoder.encode(r)
   XCTAssertEqual(lData, rData, file: (file), line: line)
